+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.1.0-CR3
 
+* Bug fix related to cancellation of queued operations when using ON_DEMAND_LOAD.
+
 ## 1.1.0-CR2
 
 * Bug fix for SEND_WITH_CALLBACK when handling operation failures.